set up tests for svelte component
