+++ src/lib/components/DataDisplay/TableDisplay.svelte
@@ -1,3 +1,33 @@
+<script context="module">
+	export function filterItems(data, searchTerm, searchableColumns) {
+		// toString here for generality
+
+		return data.filter((item) =>
+			searchableColumns.some((column) => item[column]?.toString().includes(searchTerm))
+		);
+	}
+
+	export function makePlaceholderText(data, searchableColumns) {
+		const numCols = Object.keys(data[0]).length;
+		let placeholderText = 'Filter ';
+
+		if (searchableColumns.length === 1) {
+			placeholderText = placeholderText + searchableColumns[0];
+		} else if (searchableColumns.length > 1 && searchableColumns.length <= numCols / 2) {
+			placeholderText = 'Filter ' + `any of ${searchableColumns.join(', ')}`;
+		} else if (searchableColumns.length > numCols / 2 && searchableColumns.length < numCols) {
+			const difference = Object.keys(data[0]).filter((key) => !searchableColumns.includes(key));
+			placeholderText = 'Filter all columns except ' + `${difference.join(', ')}`;
+		} else if (searchableColumns.length === numCols) {
+			placeholderText = 'Filter all columns';
+		} else {
+			placeholderText = 'Filter disabled';
+		}
+
+		return placeholderText;
+	}
+</script>
+
 <script>
 	// @ts-nocheck
 
@@ -6,8 +36,7 @@
 	import TableCell from '$lib/components/DataDisplay/TableElements/TableCell.svelte';
 	import TableHeader from '$lib/components/DataDisplay/TableElements/TableHeader.svelte';
 
-	import { filterItems, makePlaceholderText } from '$lib/components/ComponentLogic/tableLogic.js';
-	// variables
+	// exported variables
 	export let data = [];
 	export let caption = '';
 	const legendcaption = 'Meaning of indicators';
@@ -20,10 +49,10 @@
 
 	export let statusColumns = [];
 	export let searchableColumns = [];
-	// functionality
+
+	// reactive statements
 	let searchTerm = '';
 
-	// toString here for generality
 	$: filteredItems = filterItems(data, searchTerm, searchableColumns);
 
 	// make the placeholdertext for the searchbar dynamic
